Fix multiple ratings bug
