+++ src/flick/show/views.py
@@ -47,7 +47,12 @@ class ShowDetail(generics.GenericAPIView):
         score = data.get("user_rating")
 
         if score:
-            show.ratings.create(score=score, rater=user)
+            existing_rating = show.ratings.filter(rater=user)
+            if existing_rating:
+                existing_rating = show.ratings.get(rater=user)
+                existing_rating.score = score
+            else:
+                show.ratings.create(score=score, rater=user)
             show.save()
 
         return success_response(self.serializer_class(show, context={"request": request}).data)
